+++ partiql-lang/src/main/kotlin/org/partiql/lang/eval/physical/PhysicalPlanCompilerImpl.kt
@@ -173,6 +173,27 @@ internal class PhysicalPlanCompilerImpl(
         }
     }
 
+    /**
+     * Compiles a [PartiqlPhysical.Expr] tree to an [Expression].
+     *
+     * Checks [Thread.interrupted] before every expression and sub-expression is compiled
+     * and throws [InterruptedException] if [Thread.interrupted] it has been set in the
+     * hope that long-running compilations may be aborted by the caller.
+     */
+    internal fun compile(expr: PartiqlPhysical.Expr, localsSize: Int): Expression {
+        val thunk = compileAstExpr(expr)
+
+        return object : Expression {
+            override fun eval(session: EvaluationSession): ExprValue {
+                val env = EvaluatorState(
+                    session = session,
+                    registers = Array(localsSize) { ExprValue.missingValue }
+                )
+                return thunk(env)
+            }
+        }
+    }
+
     override fun convert(expr: PartiqlPhysical.Expr): PhysicalPlanThunk = this.compileAstExpr(expr)
 
     /**
@@ -183,8 +204,8 @@ internal class PhysicalPlanCompilerImpl(
     private fun compileAstStatement(ast: PartiqlPhysical.Statement): PhysicalPlanThunk {
         return when (ast) {
             is PartiqlPhysical.Statement.Query -> compileAstExpr(ast.expr)
-            is PartiqlPhysical.Statement.DmlQuery -> compileAstExpr(ast.expr)
             is PartiqlPhysical.Statement.Exec -> compileExec(ast)
+            is PartiqlPhysical.Statement.Dml,
             is PartiqlPhysical.Statement.Explain -> {
                 val value = ExprValue.newBoolean(true)
                 thunkFactory.thunkEnv(emptyMetaContainer()) { value }